# public-transport-rest-contract/src/main/resources/openapi.yml
openapi: 3.0.1
info:
  title: Public transport ticket endpoint
  description: This is an application which is simulate a public transport ticket handling.
    This application based on Budapest public transport (BKK) ticket types.
  contact:
    email: karoly.froemel@gmail.com
  version: 2.0.0
servers:
  - url: http://localhost:8081
paths:
  /ticketType:
    get:
      tags:
        - Ticket type
      operationId: getTicketTypes
      summary: Get all ticket types
      description: Get all types of tickets to show it for the user.
      responses:
        200:
          description: Get all types of tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TicketType'
        401:
          $ref: '#/components/responses/401Unauthorized'
        403:
          $ref: '#/components/responses/403PermissionDenied'
        404:
          $ref: '#/components/responses/404ContentNotFound'
        500:
          $ref: '#/components/responses/500ServerError'

    post:
      tags:
        - Ticket type
      operationId: createNewTicketType
      summary: Add a new ticket type.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketType"
      responses:
        201:
          description: Ticket type succesfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketType'
        400:
          $ref: '#/components/responses/400BadRequest'
        401:
          $ref: '#/components/responses/401Unauthorized'
        403:
          $ref: '#/components/responses/403PermissionDenied'
        500:
          $ref: '#/components/responses/500ServerError'

    put:
      tags:
        - Ticket type
      operationId: updateTicketType
      summary: Update a ticket type.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketType'
      responses:
        204:
          description: Successfully updated.
        400:
          $ref: '#/components/responses/400BadRequest'
        401:
          $ref: '#/components/responses/401Unauthorized'
        403:
          $ref: '#/components/responses/403PermissionDenied'
        404:
          $ref: '#/components/responses/404ContentNotFound'
        500:
          $ref: '#/components/responses/500ServerError'
  /ticketType/{id}:
    get:
      tags:
        - Ticket type
      operationId: getTicketTypeById
      summary: Get ticket type by id.
      parameters:
        - name: id
          in: path
          required: true
          description: ticket type id
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Succesfully get ticket by id.
          content:
            application/json:
              schema:
                $ref:  "#/components/schemas/TicketType"
        401:
          $ref: '#/components/responses/401Unauthorized'
        403:
          $ref: '#/components/responses/403PermissionDenied'
        404:
          $ref: '#/components/responses/404ContentNotFound'
        500:
          $ref: '#/components/responses/500ServerError'

    delete:
      tags:
        - Ticket type
      operationId: deleteTicketTypeById
      summary: Delete ticket type by id.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        204:
          description: Successfully deleted.
        401:
          $ref: '#/components/responses/401Unauthorized'
        403:
          $ref: '#/components/responses/403PermissionDenied'
        404:
          $ref: '#/components/responses/404ContentNotFound'
        500:
          $ref: '#/components/responses/500ServerError'
  /ticketType/search:
    post:
      tags:
        - Ticket type
      operationId: searchTicketType
      summary: Search ticket type.
      parameters:
        - $ref: '#/components/parameters/X-Page'
        - $ref: '#/components/parameters/X-Size'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketTypeSearchRequest'
      responses:
        206:
          description: 'Successful request'
          headers:
            X-Page:
              $ref: '#/components/headers/X-Page'
            X-Size:
              $ref: '#/components/headers/X-Size'
            X-Total-Pages:
              $ref: '#/components/headers/X-Total-Pages'
            X-Total-Size:
              $ref: '#/components/headers/X-Total-Size'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TicketType'
        400:
          $ref: '#/components/responses/400BadRequest'
        403:
          $ref: '#/components/responses/403PermissionDenied'
        500:
          $ref: '#/components/responses/500ServerError'
  /ticketType/searchLambda:
    post:
      tags:
        - Ticket type
      operationId: searchTicketTypeLambda
      summary: Search ticket type by lambda expression.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketTypeSearchRequest'
      responses:
        200:
          description: 'Successful request'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TicketType'
        400:
          $ref: '#/components/responses/400BadRequest'
        403:
          $ref: '#/components/responses/403PermissionDenied'
        500:
          $ref: '#/components/responses/500ServerError'
...
components:
  schemas:
    TicketType:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 50
        description:
          type: string
          maxLength: 250
        price:
          type: integer
          minimum: 1
        expirationTime:
          type: integer
          minimum: 1
      required:
        - name
        - description
        - price
        - expirationTime
    TicketTypeSearchRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 50
          nullable: false
          description: 'Name of ticket type'
        description:
          type: string
          maxLength: 250
          description: 'Description of ticket type'
        sortBy:
          type: string
          nullable: false
          default: 'name'
          description: 'Sort by this column'
        sortOrder:
          type: string
          enum:
            - ASC
            - DESC
          default: 'ASC'
          description: 'Sort Order'
...
    Error:
      title: "Error object"
      type: object
      properties:
        errorCode:
          type: string
          nullable: false
          description: "Error code"
        message:
          type: string
          nullable: false
          description: "Error message"
        detailedMessage:
          type: string
          nullable: false
          description: "Detailed error message"
      required:
        - "errorCode"
  parameters:
    X-Page:
      name: X-Page
      in: header
      description: "Requested page number"
      required: true
      schema:
        type: integer
        format: int64
    X-Size:
      name: X-Size
      in: header
      description: "Requested page size"
      required: true
      schema:
        type: integer
        format: int64
  headers:
    X-Page:
      required: true
      schema:
        type: integer
        format: int64
      description: "Number of requested pages"
    X-Size:
      required: true
      schema:
        type: integer
        format: int64
      description: "Max page size"
    X-Total-Pages:
      required: true
      schema:
        type: integer
        format: int64
      description: "Number of all requestable pages"
    X-Total-Size:
      required: true
      schema:
        type: integer
        format: int64
      description: "Number of all requestable element"
  responses:
    400BadRequest:
      description: "Bad request."
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    401Unauthorized:
      description: "Authentication is required"
    403PermissionDenied:
      description: "User do not have permission to perform the action."
    404ContentNotFound:
      description: "Content not found."
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    500ServerError:
      description: "Server side error."
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'